add PostController suite

+++ Rest Assured Testing - WeAre Project/src/test/java/weare/api/testing/post/CreatePostTest.java
@@ -62,8 +62,11 @@ public class CreatePostTest extends BaseTestSetup {
         editPost = createdPost;
         postId = createdPost.postId;
 
-        PostController.deletePost(cookies, createdPost.postId);
-        System.out.println("Post with Id" + " " + postId + " " + "Deleted successfully.");
+        if(!isDeletedPost){
+            PostController.deletePost(cookies, createdPost.postId);
+            System.out.println("Post with Id" + " " + postId + " " + "Deleted successfully.");
+            isDeletedPost = true;
+        }
     }
 }
